pull out ConstantFoldingTests
